+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- [#611](https://github.com/loop-payments/prisma-lint/issues/611) Add `nativeType` option to `require-field-type` rule to additionally enforce native DB types if needed.
+
 ## 0.8.0 (2025-01-17)
 
 - Add new rule `field-name-grammatical-number`.